+++ example/lib/feature/button_page.dart
@@ -1,5 +1,6 @@
 import 'package:express_widget/express_widget.dart';
 import 'package:flutter/material.dart';
+import 'package:ionicons/ionicons.dart';
 
 class ButtonPage extends StatelessWidget {
   const ButtonPage({super.key});
@@ -15,7 +16,7 @@ class ButtonPage extends StatelessWidget {
             children: [
               SizedBox(
                 width: MediaQuery.of(context).size.width,
-                child: AppButton(
+                child: ExpressButton(
                   appButtonStyle: AppButtonStyle.rectangle,
                   onPressed: () {
                     debugPrint("on pressed");
@@ -27,7 +28,7 @@ class ButtonPage extends StatelessWidget {
               ),
               SizedBox(
                 width: MediaQuery.of(context).size.width,
-                child: AppButton(
+                child: ExpressButton(
                   appButtonStyle: AppButtonStyle.rectangle,
                   onPressed: () {
                     debugPrint("on pressed");
@@ -40,7 +41,7 @@ class ButtonPage extends StatelessWidget {
               ),
               SizedBox(
                 width: MediaQuery.of(context).size.width,
-                child: AppButton(
+                child: ExpressButton(
                   appButtonStyle: AppButtonStyle.rectangle,
                   onPressed: () {
                     debugPrint("on pressed");
@@ -49,6 +50,85 @@ class ButtonPage extends StatelessWidget {
                   textColor: Colors.white,
                   isEnabled: false,
                   backgroundColor: Theme.of(context).primaryColor,
+                ),
+              ),
+              SizedBox(
+                width: MediaQuery.of(context).size.width,
+                child: ExpressOutlineButton(
+                  appButtonStyle: AppButtonStyle.rectangle,
+                  onPressed: () {
+                    debugPrint("on pressed");
+                  },
+                  text: "Hello",
+                  textColor: Colors.white,
+                  backgroundColor: Theme.of(context).primaryColor,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      const Icon(Ionicons.logo_google),
+                      Container(width: 20),
+                      const Text(
+                        "Continue with Google",
+                        style: TextStyle(
+                          color: appColorBlack,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+              SizedBox(
+                width: MediaQuery.of(context).size.width,
+                child: ExpressOutlineButton(
+                  appButtonStyle: AppButtonStyle.rectangle,
+                  onPressed: () {
+                    debugPrint("on pressed");
+                  },
+                  text: "Hello",
+                  textColor: Colors.white,
+                  backgroundColor: Theme.of(context).primaryColor,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      const Icon(Ionicons.logo_apple),
+                      Container(width: 20),
+                      const Text(
+                        "Continue with Apple",
+                        style: TextStyle(
+                          color: appColorBlack,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+              SizedBox(
+                width: MediaQuery.of(context).size.width,
+                child: ExpressOutlineButton(
+                  appButtonStyle: AppButtonStyle.rectangle,
+                  onPressed: () {
+                    debugPrint("on pressed");
+                  },
+                  text: "Hello",
+                  textColor: Colors.white,
+                  backgroundColor: Theme.of(context).primaryColor,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    mainAxisSize: MainAxisSize.max,
+                    children: [
+                      const Icon(Ionicons.logo_facebook),
+                      Container(width: 20),
+                      const Text(
+                        "Continue with Facebook",
+                        style: TextStyle(
+                          color: appColorBlack,
+                        ),
+                      ),
+                    ],
+                  ),
                 ),
               ),
             ],