+++ lib/src/style/app_color.dart
@@ -2,4 +2,5 @@ import 'package:flutter/material.dart';
 
 const appColorBlack = Color(0xff101010);
 const appColorGray = Color(0xfff4f5f7);
+const appColorWhite = Color(0xffffffff);
 const appColorDarkGray = Color(0xffc6c6c6);